requirements.txt
python-pptx>=1.0.0
matplotlib>=3.8.0
Pillow>=10.0.0
